+++ Core/KingdomInheritanceSpatial.CaptureResult.cs
@@ -1,8 +1,3 @@
-using System;
-using System.Collections.Generic;
-using XRL;
-using XRL.World;
-
 namespace ThousandAndFirst
 {
 	internal enum KingdomInheritanceSpatialCaptureResult